docs: improve/simplify example for `providers` guide
