--- aio/content/examples/providers/src/app/core/core.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-
-import { UserService } from './user.service';
-
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule
-  ],
-  declarations: [],
-  providers: [UserService]
-})
-export class CoreModule { }
