--- aio/content/examples/providers/src/app/core/mock-users.ts
@@ -1,14 +1,0 @@
-import { User } from './user';
-
-export const USERS: User[] = [
-  { id: 1, name: 'Maria' },
-  { id: 2, name: 'Alex' },
-  { id: 3, name: 'Chuntao' },
-  { id: 4, name: 'Béatrice' },
-  { id: 5, name: 'Sarah' },
-  { id: 6, name: 'Andrés' },
-  { id: 7, name: 'Abdul' },
-  { id: 8, name: 'Pierre' },
-  { id: 9, name: 'Jiao' },
-  { id: 10, name: 'Seth' }
-];
