--- aio/content/examples/providers/src/app/core/user.service.ts
@@ -1,15 +1,0 @@
-import { Injectable } from '@angular/core';
-
-import { User } from './user';
-import { USERS } from './mock-users';
-
-@Injectable()
-export class UserService {
-
-  constructor() { }
-
-
-  getUsers(): Promise<User[]> {
-    return Promise.resolve(USERS);
-  }
-}
