--- aio/content/examples/providers/src/app/core/user.ts
@@ -1,4 +1,0 @@
-export class User {
-  id: number;
-  name: string;
-}
